moved to other files so people can work on main.c
